Round values when using masks and factors or offsets at the same time
When using the bool type in combination with a factor, the outcome is not as expected.

I use this to have an on/off contact which drives a pneumatic cilinder with length of 0.25m.

Factor in this case will be 4. 
During planning the value will be between 0 and 1, but due to rounding errors it never get's to 1, always stays at 0.9999...

With this rounding, we now can use this to round the 0.9999 to a 1.