Transition from ament_target_dependencies to target_link_libraries
Since kilted, the use of ament_target_dependencies has been deprecated.
This commit replaces all uses of the macro with target_link_libraries as
suggested in the deprecation warning, and does some additional steps in
for the ethercat_interface package to ensure exports are done in the
"modern" target-based fashion.