Update the way the yaml libraries are linked
When using the yaml_cpp_vendor package, the correct way to link agains
the yaml libraries is to first use:
  find_package(yaml_cpp_vendor REQUIRED)
  find_package(PkgConfig REQUIRED)
  pkg_check_modules(YAML_CPP REQUIRED yaml-cpp)
to find the concrete yaml vendor package, and then use the YAML_CPP_*
CMake variables to get the link targets, include libraries, and cmake
options. This commit updates the relevant cmake files to comply with
this pattern.